Code clean-up: move shutter-related tasks to separate file.
