+++ utils.c
@@ -13,12 +13,9 @@
 #include "debug.h"
 
 #include "firmware.h"
-#include "firmware/camera.h"
+//#include "firmware/camera.h"
 
 #include "utils.h"
-
-void lock_sutter     (void);
-void wait_for_shutter(void);
 
 void display_float(char *dest, float value);
 
@@ -182,93 +179,6 @@ void start_debug_mode() {
 	printf("\n");
 
 	beep();
-}
-
-#if 0
-// this is a disassembled version of eventproc_release()
-int shutter_release_disasm() {
-
-	extern char * aRelSem;
-
-	if (hRelSem == 0) {
-		hRelSem = CreateBinarySemaphore(aRelSem, 0);
-	}
-
-	SendToIntercom(IC_RELEASE, 0, 0);
-	SendToIntercom(0x6D, 1, 1); // set burst counter
-
-	TakeSemaphore(hRelSem, 30000);
-	DeleteSemaphore(hRelSem);
-	hRelSem = 0;
-
-	SleepTask(EVENT_WAIT); // we added this
-	return 0;
-}
-#endif
-
-void lock_sutter(void) {
-	shutter_lock = TRUE;
-}
-
-void wait_for_shutter(void) {
-	while (shutter_lock)
-		SleepTask(RELEASE_WAIT);
-}
-
-void wait_for_camera() {
-	while (! able_to_release())
-		SleepTask(RELEASE_WAIT);
-}
-
-int shutter_release() {
-	wait_for_camera();
-	lock_sutter    ();
-
-	int result = press_button(IC_BUTTON_FULL_SHUTTER);
-
-	if (DPData.drive == DRIVE_MODE_TIMER)
-		SleepTask(SELF_TIMER_MS);
-
-	wait_for_shutter();
-
-	return result;
-}
-
-int shutter_release_bulb(int time) {
-	static int first = TRUE;
-
-	int  button;
-	long delay;
-
-	int shutter_lag, mirror_lag;
-
-	if (first) {
-		first = FALSE;
-		shutter_lag = SHUTTER_LAG_1ST;
-		mirror_lag  = MIRROR_LAG_1ST;
-	} else {
-		shutter_lag = SHUTTER_LAG_2ND;
-		mirror_lag  = MIRROR_LAG_2ND;
-	}
-
-	if (DPData.drive == DRIVE_MODE_TIMER) {
-		button = IC_BUTTON_FULL_SHUTTER;
-		delay  = time + shutter_lag + mirror_lag;
-	} else {
-		button = IC_BUTTON_HALF_SHUTTER;
-		delay  = time + shutter_lag;
-	}
-
-	wait_for_camera();
-	lock_sutter    ();
-
-	press_button(button);
-	SleepTask   (delay);
-	press_button(button);
-
-	wait_for_shutter();
-
-	return 0;
 }
 
 int print_icu_info() {
